pipeline_layer/requirements.txt
scikit-learn
PyYAML
pandas
numpy
paho-mqtt